requirements.txt
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.0
defusedxml==0.6.0
Django==2.1.8
django-autocomplete-light==3.3.5
django-ckeditor==5.7.1
django-js-asset==1.2.2
django-rest-framework==0.1.0
djangorestframework==3.9.4
entrypoints==0.3
idna==2.8
image==1.5.27
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
pandocfilters==1.4.2
parso==0.4.0
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
Pygments==2.4.0
pyrsistent==0.15.1
python-dateutil==2.8.0
pytz==2019.1
pywinpty==0.5.5
pyzmq==18.0.1
qtconsole==4.4.4
requests==2.22.0
Send2Trash==1.5.0
six==1.12.0
some-package==0.1
sqlparse==0.3.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.25.2
virtualenv==16.6.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
